Remove test_if_from as a necessary method
Turns out that `test_if_from` is not used by the driver directly.
It is called after instantiating the parser class, and the constructor raises
an exception if `test_if_from` fails. So, there really isn't a point
in having `test_if_from` as a separate operation. It never gets
called by classes that use this parser because the parser fails
to instantiate before you call this function.

I had originally intended this method as a static operation to check
whether a directory of files was likely generated by a certain code,
but it seems now that just having the constructor for a parser fail
if the files are not as expected serves the same purpose.
Even if we implemented these as static methods, I doubt we gain much.
We have to do complicated logic to determine whether a directory is a
certain type (see PWSCF). Why bother calling that logic twice: once for
checking if I could make the parser, and again when making the parser
(as you'll have to determine which files are the inputs/outputs
to both check if a directory contains these files and extracting
metadata from those classes).

Consequently, I've moved the logic from `test_if_from` inside the
constructors and simplified the driver.

Sorry for the long PR message. Does this make sense?