Added ability to filter which files a parser can read
I did this by introducing a kwarg to the base DFTParser
class that defines a list of files a parser is allowed to read (files)
If `None`, the parser can read all of the files in the input directory.
The list of acceptable files (either user-provided, or assembled by
listing the directory contents) is stored in a private variable
`self._files`, which mirrors the base directory (`self._directory`).

I have also changed the implementation of the VASP and PWSCF parser
to support this selective file reading, and the driver to pass
information on restricted files to the parsers.

Do you think this is ugly?

Broken tests: PWSCF VS2 has two inputs files, I'll fix that later